@@ -1,29 +1,28 @@
 from abc import ABC, abstractmethod
 from collections.abc import Sequence
+from dataclasses import dataclass
 from functools import cached_property
 from pathlib import Path
-from typing import TYPE_CHECKING, Generic, TypeVar
-
-from pydantic import BaseModel
+from typing import TYPE_CHECKING
 
 if TYPE_CHECKING:
     from ._base import BaseStorage
 
-_N = TypeVar("_N", bound="TreeNode")
-_S = TypeVar("_S", bound="BaseStorage")
+    """Base class for tree nodes in storage trees."""
 
 
-class TreeNode(BaseModel):
+@dataclass
+class TreeNode:
     """Base class for tree nodes in storage trees."""
 
 
 # TODO(@mghali): add total_dirs and total_files like unix cli tree at the end of output
-class StorageTree(ABC, Generic[_S, _N]):
+class StorageTree[S: BaseStorage, N: TreeNode](ABC):
     """Abstract base class for storage tree structures."""
 
-    _storage: _S
+    _storage: S
 
-    def __init__(self, storage: _S) -> None:
+    def __init__(self, storage: S) -> None:
         """Initialize storage tree."""
         self._storage = storage
 
@@ -36,16 +35,16 @@ class StorageTree(ABC, Generic[_S, _N]):
     def levels(self) -> int: ...
 
     @abstractmethod
-    def next(self) -> _N: ...
+    def next(self) -> N: ...
 
     @abstractmethod
-    def previous(self) -> _N: ...
+    def previous(self) -> N: ...
 
     @abstractmethod
     def draw(self) -> str: ...
 
     @abstractmethod
-    def search(self, pattern: str) -> Sequence[_N]: ...
+    def search(self, pattern: str) -> Sequence[N]: ...
 
     # TODO(@mghali): should i implement those, also include them as algorithm selection in
     # search method? do they return dictionary of nodes? check anytree lib
