@@ -1,9 +1,9 @@
+import datetime as dt
 from collections.abc import Iterator, Sequence
 from contextlib import contextmanager
-from datetime import datetime as dt
 from pathlib import Path
 from types import TracebackType
-from typing import Any, Literal, overload
+from typing import Any, Literal, Self, overload
 
 import magic
 from azure.storage.blob import ContentSettings
@@ -15,11 +15,11 @@ from azure.storage.filedatalake import (
 )
 from loguru import logger
 from pydantic import BaseModel
-from typing_extensions import Self
 
 from storix.sandbox import PathSandboxable, SandboxedPathHandler
+from storix.security import SAS_EXPIRY_SECONDS, SAS_PERMISSIONS, Permissions
 from storix.settings import settings
-from storix.typing import PathLike
+from storix.typing import StrPathLike
 
 from ._base import BaseStorage
 
@@ -29,8 +29,8 @@ class FileProperties(BaseModel):
 
     name: str
     hdi_isfolder: bool = False
-    last_modified: dt
-    creation_time: dt
+    last_modified: dt.datetime
+    creation_time: dt.datetime
 
     model_config = {"from_attributes": True}
 
@@ -39,24 +39,24 @@ class AzureDataLake(BaseStorage):
     """Azure Data Lake Storage Gen2 implementation."""
 
     __slots__ = (
-        "__filesystem",
-        "__service_client",
         "_current_path",
+        "_filesystem",
         "_home",
         "_min_depth",
         "_sandbox",
+        "_service_client",
     )
 
     _sandbox: PathSandboxable | None
-    __service_client: DataLakeServiceClient
-    __filesystem: FileSystemClient
+    _service_client: DataLakeServiceClient
+    _filesystem: FileSystemClient
     _home: Path
     _current_path: Path
     _min_depth: Path
 
     def __init__(
         self,
-        initialpath: PathLike | None = None,
+        initialpath: StrPathLike | None = None,
         container_name: str = str(settings.ADLSG2_CONTAINER_NAME),
         adlsg2_account_name: str | None = settings.ADLSG2_ACCOUNT_NAME,
         adlsg2_token: str | None = settings.ADLSG2_TOKEN,
@@ -102,11 +102,11 @@ class AzureDataLake(BaseStorage):
             "ADLSg2 account name and authentication token are required"
         )
 
-        self.__service_client = self._get_service_client(
+        self._service_client = self._get_service_client(
             adlsg2_account_name, adlsg2_token
         )
-        self.__filesystem = self._init_filesystem(
-            self.__service_client, str(container_name)
+        self._filesystem = self._init_filesystem(
+            self._service_client, str(container_name)
         )
 
         super().__init__(
@@ -128,12 +128,12 @@ class AzureDataLake(BaseStorage):
 
     # TODO(@mghali): convert the return type to dict[str, str] or Tree DS
     # so that its O(1) from the ui-side to access
-    def tree(self, path: PathLike | None = None, *, abs: bool = False) -> list[Path]:
+    def tree(self, path: StrPathLike | None = None, *, abs: bool = False) -> list[Path]:
         """Return a tree view of files and directories starting at path."""
         path = self._topath(path)
         self._ensure_exist(path)
 
-        all = self.__filesystem.get_paths(path=str(path), recursive=True)
+        all = self._filesystem.get_paths(path=str(path), recursive=True)
         paths: list[Path] = [self._topath(f.name) for f in all]
 
         if self._sandbox:
@@ -144,23 +144,23 @@ class AzureDataLake(BaseStorage):
     @overload
     def ls(
         self,
-        path: PathLike | None = None,
+        path: StrPathLike | None = None,
         *,
         abs: Literal[False] = False,
         all: bool = True,
     ) -> list[str]: ...
     @overload
     def ls(
-        self, path: PathLike | None = None, *, abs: Literal[True], all: bool = True
+        self, path: StrPathLike | None = None, *, abs: Literal[True], all: bool = True
     ) -> list[Path]: ...
     def ls(
-        self, path: PathLike | None = None, *, abs: bool = False, all: bool = True
+        self, path: StrPathLike | None = None, *, abs: bool = False, all: bool = True
     ) -> Sequence[Path | str]:
         """List all items at the given path as Path or str objects."""
         path = self._topath(path)
         self._ensure_exist(path)
 
-        items = self.__filesystem.get_paths(path=str(path), recursive=False)
+        items = self._filesystem.get_paths(path=str(path), recursive=False)
         paths: Sequence[Path] = [self.home / f.name for f in items]
 
         if not all:
@@ -171,18 +171,18 @@ class AzureDataLake(BaseStorage):
 
         return paths
 
-    def mkdir(self, path: PathLike, *, parents: bool = False) -> None:
+    def mkdir(self, path: StrPathLike, *, parents: bool = False) -> None:
         """Create a directory at the given path."""
         path = self._topath(path)
         # TODO(mghalix): add parents logic
-        self.__filesystem.create_directory(str(path))
+        self._filesystem.create_directory(str(path))
 
-    def isdir(self, path: PathLike) -> bool:
+    def isdir(self, path: StrPathLike) -> bool:
         """Check if the given path is a directory."""
         stats = self.stat(path)
         return stats.hdi_isfolder
 
-    def stat(self, path: PathLike) -> FileProperties:
+    def stat(self, path: StrPathLike) -> FileProperties:
         """Return stat information for the given path."""
         path = self._topath(path)
         self._ensure_exist(path)
@@ -196,13 +196,13 @@ class AzureDataLake(BaseStorage):
 
             return FileProperties.model_validate(dict(**props, **metadata))
 
-    def isfile(self, path: PathLike) -> bool:
+    def isfile(self, path: StrPathLike) -> bool:
         """Check if the given path is a file."""
         stats = self.stat(path)
         return not stats.hdi_isfolder
 
     # TODO(@mghali): add a confirm override option
-    def mv(self, source: PathLike, destination: PathLike) -> None:
+    def mv(self, source: StrPathLike, destination: StrPathLike) -> None:
         """Move a file or directory to a new location."""
         source = self._topath(source)
         destination = self._topath(destination)
@@ -221,7 +221,7 @@ class AzureDataLake(BaseStorage):
         self.touch(dest, data)
         self.rm(source)
 
-    def cd(self, path: PathLike | None = None) -> Self:
+    def cd(self, path: StrPathLike | None = None) -> Self:
         """Change the current working directory."""
         if path is None:
             path = self.home
@@ -240,7 +240,7 @@ class AzureDataLake(BaseStorage):
         self._current_path = path
         return self
 
-    def rm(self, path: PathLike) -> bool:
+    def rm(self, path: StrPathLike) -> bool:
         """Remove a file at the given path."""
         path = self._topath(path)
 
@@ -261,7 +261,7 @@ class AzureDataLake(BaseStorage):
 
         return True
 
-    def rmdir(self, path: PathLike, recursive: bool = False) -> bool:
+    def rmdir(self, path: StrPathLike, recursive: bool = False) -> bool:
         """Remove a directory at the given path."""
         path = self._topath(path)
 
@@ -280,7 +280,7 @@ class AzureDataLake(BaseStorage):
 
         return True
 
-    def touch(self, path: PathLike | None, data: Any | None = None) -> bool:
+    def touch(self, path: StrPathLike | None, data: Any | None = None) -> bool:
         """Create a file at the given path, optionally writing data."""
         path = self._topath(path)
 
@@ -299,7 +299,7 @@ class AzureDataLake(BaseStorage):
 
         return True
 
-    def cat(self, path: PathLike) -> bytes:
+    def cat(self, path: StrPathLike) -> bytes:
         """Read the contents of a file as bytes."""
         path = self._topath(path)
         self._ensure_exist(path)
@@ -314,7 +314,7 @@ class AzureDataLake(BaseStorage):
 
         return blob
 
-    def exists(self, path: PathLike) -> bool:
+    def exists(self, path: StrPathLike) -> bool:
         """Check if the given path exists."""
         path = self._topath(path)
 
@@ -331,7 +331,7 @@ class AzureDataLake(BaseStorage):
             except Exception:
                 return False
 
-    def cp(self, source: PathLike, destination: PathLike) -> None:
+    def cp(self, source: StrPathLike, destination: StrPathLike) -> None:
         """Copy a file or directory to a new location."""
         source = self._topath(source)
         destination = self._topath(destination)
@@ -344,26 +344,87 @@ class AzureDataLake(BaseStorage):
         # TODO(@mghali): copy tree
         raise NotImplementedError
 
-    def du(self, path: PathLike | None = None, *, human_readable: bool = True) -> Any:
+    def du(
+        self, path: StrPathLike | None = None, *, human_readable: bool = True
+    ) -> Any:
         """Return disk usage statistics for the given path."""
         raise NotImplementedError
 
     def close(self) -> None:
         """Close the Azure Data Lake storage client."""
-        self.__filesystem.close()
-        self.__service_client.close()
+        self._filesystem.close()
+        self._service_client.close()
 
     @contextmanager
-    def _get_file_client(self, filepath: PathLike) -> Iterator[DataLakeFileClient]:
+    def _get_file_client(self, filepath: StrPathLike) -> Iterator[DataLakeFileClient]:
         filepath = self._topath(filepath)
-        with self.__filesystem.get_file_client(str(filepath)) as client:
+        with self._filesystem.get_file_client(str(filepath)) as client:
             yield client
 
     @contextmanager
-    def _get_dir_client(self, dirpath: PathLike) -> Iterator[DataLakeDirectoryClient]:
+    def _get_dir_client(
+        self, dirpath: StrPathLike
+    ) -> Iterator[DataLakeDirectoryClient]:
         dirpath = self._topath(dirpath)
-        with self.__filesystem.get_directory_client(str(dirpath)) as client:
+        with self._filesystem.get_directory_client(str(dirpath)) as client:
             yield client
+
+    def make_url(
+        self, path: StrPathLike, *, astype: Literal["data_url", "sas"] = "sas"
+    ) -> str:
+        """Generate a url for a path."""
+        if astype == "sas":
+            return self._generate_sas_url(
+                path, expires_in=SAS_EXPIRY_SECONDS, permissions=SAS_PERMISSIONS
+            )
+        return super().make_url(path, astype=astype)
+
+    def _generate_sas_url(
+        self,
+        path: StrPathLike,
+        *,
+        expires_in: int = 3600,
+        permissions: frozenset[Permissions] = frozenset({Permissions.READ}),
+    ) -> str:
+        from azure.storage.filedatalake import FileSasPermissions, generate_file_sas
+
+        from storix.utils import craft_adlsg2_url_sas
+
+        path = self._topath(path)
+        self._ensure_exist(path)
+
+        if self.isdir(path):
+            raise ValueError("cannot generate a sas token for a directory")
+
+        fs = self._filesystem
+        account_name: str = str(fs.account_name)
+        container: str = str(fs.file_system_name)
+        credential: str = str(fs.credential.account_key)
+
+        expiry = dt.datetime.now(dt.UTC) + dt.timedelta(seconds=expires_in)
+        file_permissions = FileSasPermissions(**dict.fromkeys(permissions, True))
+
+        directory: str = str(path.parent).lstrip("/")
+        filename: str = path.parts[-1]
+
+        # pure local crypto op no i/o
+        token = generate_file_sas(
+            account_name=account_name,
+            file_system_name=container,
+            credential=credential,
+            directory_name=directory,
+            file_name=filename,
+            permission=file_permissions,
+            expiry=expiry,
+        )
+
+        return craft_adlsg2_url_sas(
+            account_name=account_name,
+            container=container,
+            directory=directory,
+            filename=filename,
+            sas_token=token,
+        )
 
     def __enter__(self) -> Self:
         """Enter the runtime context related to this object."""
